pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "strata"
version = "0.1.0"
description = "Composable infrastructure templates for Docker Compose"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "jinja2>=3.1"
]

[project.scripts]
strata = "strata.cli:cli"
